% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ioCal}
\alias{ioCal}
\alias{readCal}
\alias{writeCal}
\title{Save and read HMM calibration to file generated by
    \code{\link[groHMM]{evaluateHMMInAnnotations}} \code{\link{data.frame}}.}
\usage{
writeCal(data, file)

readCal(file)
}
\arguments{
\item{data}{A data.frame obtained from
\code{\link[groHMM]{evaluateHMMInAnnotations}}.}

\item{file}{A string path.}
}
\examples{
\dontrun{comparison <- groHMM::evaluateHMMInAnnotations(hmm$transcripts, consensus)}
\dontrun{hmm <- calibrate(reads=reads, LtProbB=LtProbB, UTS=UTS)}
fileCal <- "cell_line.grohmm.cal"
writeCal(ans, fileCal)
## Now that we have a calibration for our cell line, we can use it in later R
## sessions:
cal <- readCal(fileCal)

}
\seealso{
\code{\link[groHMM]{evaluateHMMInAnnotations}} for
}

