% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{readBams}
\alias{readBams}
\title{Read BAM files as \code{\linkS4class[GenomicRanges]{GRanges}} object.

Also sanitizes invalid chromosomes using \code{\link{limitChromosomes}}.}
\usage{
readBams(bams, filter = TRUE)
}
\arguments{
\item{bams}{A string vector of BAM file paths.}
}
\value{
\code{\linkS4class[GenomicRanges]{GRanges}} S4 object combining all
    \code{bams}.
}
\examples{
gr <- readBams("foo.bam")
gr <- readBams(c("foo.bam", "baz/foo.bam"))
}
\seealso{
\code{\link{limitChromosomes}} for sanitization, and
    \code{\linkS4class[GenomicRanges]{GRanges}} for returned value.
}

