% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runon.R
\docType{package}
\name{runon-package}
\alias{runon}
\alias{runon-package}
\title{Analyze putative genes from nuclear run on experiments.}
\description{
runon allows you to process GRO-seq and PRO-seq data to infer gene
boundaries using an Hidden Markov Model (HMM).
}
\details{
For more details, see the package vignette:
\code{browseVignettes(package = "runon")}
}

